+++ prqlc/prqlc-parser/src/test.rs
@@ -1,29 +1,32 @@
+use chumsky::Parser;
 use insta::{assert_debug_snapshot, assert_yaml_snapshot};
-use itertools::Itertools;
 
-use crate::error::Error;
-use crate::lexer;
-use crate::parser;
-use crate::parser::pr::{Expr, FuncCall, Stmt};
+use crate::parser::pr::Stmt;
+use crate::parser::prepare_stream;
+use crate::parser::stmt;
+use crate::{error::Error, lexer::lr::TokenKind, parser::perror::PError};
 
-/// Helper that does not track source_ids
-fn parse_single(source: &str) -> Result<Vec<Stmt>, Vec<Error>> {
-    let tokens = lexer::lex_source(source)?;
+/// Parse source code based on the supplied parser.
+///
+/// Use this to test any parser!
+pub(crate) fn parse_with_parser<O>(
+    source: &str,
+    parser: impl Parser<TokenKind, O, Error = PError>,
+) -> Result<O, Vec<Error>> {
+    let tokens = crate::lexer::lex_source(source)?;
+    let stream = prepare_stream(tokens.0.into_iter(), source, 0);
 
-    let (ast, parse_errors) = parser::parse_lr_to_pr(source, 0, tokens.0);
+    let (ast, parse_errors) = parser.parse_recovery(stream);
 
     if !parse_errors.is_empty() {
-        return Err(parse_errors);
+        return Err(parse_errors.into_iter().map(|e| e.into()).collect());
     }
-    Ok(ast.unwrap_or_default())
+    Ok(ast.unwrap())
 }
 
-fn parse_expr(source: &str) -> Result<Expr, Vec<Error>> {
-    let source = format!("let result = ({source}\n)");
-
-    let stmts = parse_single(&source)?;
-    let stmt = stmts.into_iter().exactly_one().unwrap();
-    Ok(*stmt.kind.into_var_def().unwrap().value.unwrap())
+/// Parse into statements
+fn parse_single(source: &str) -> Result<Vec<Stmt>, Vec<Error>> {
+    parse_with_parser(source, stmt::source())
 }
 
 #[test]
@@ -186,674 +189,6 @@ fn test_take() {
 }
 
 #[test]
-fn test_ranges() {
-    assert_yaml_snapshot!(parse_expr(r#"3..5"#).unwrap(), @r###"
-    ---
-    Range:
-      start:
-        Literal:
-          Integer: 3
-        span: "0:14-15"
-      end:
-        Literal:
-          Integer: 5
-        span: "0:17-18"
-    span: "0:13-20"
-    "###);
-
-    assert_yaml_snapshot!(parse_expr(r#"-2..-5"#).unwrap(), @r###"
-    ---
-    Range:
-      start:
-        Unary:
-          op: Neg
-          expr:
-            Literal:
-              Integer: 2
-            span: "0:15-16"
-        span: "0:14-16"
-      end:
-        Unary:
-          op: Neg
-          expr:
-            Literal:
-              Integer: 5
-            span: "0:19-20"
-        span: "0:18-20"
-    span: "0:13-22"
-    "###);
-
-    assert_yaml_snapshot!(parse_expr(r#"(-2..(-5 | abs))"#).unwrap(), @r###"
-    ---
-    Range:
-      start:
-        Unary:
-          op: Neg
-          expr:
-            Literal:
-              Integer: 2
-            span: "0:16-17"
-        span: "0:15-17"
-      end:
-        Pipeline:
-          exprs:
-            - Unary:
-                op: Neg
-                expr:
-                  Literal:
-                    Integer: 5
-                  span: "0:21-22"
-              span: "0:20-22"
-            - Ident: abs
-              span: "0:25-28"
-        span: "0:20-28"
-    span: "0:13-32"
-    "###);
-
-    assert_yaml_snapshot!(parse_expr(r#"(2 + 5)..'a'"#).unwrap(), @r###"
-    ---
-    Range:
-      start:
-        Binary:
-          left:
-            Literal:
-              Integer: 2
-            span: "0:15-16"
-          op: Add
-          right:
-            Literal:
-              Integer: 5
-            span: "0:19-20"
-        span: "0:15-20"
-      end:
-        Literal:
-          String: a
-        span: "0:23-26"
-    span: "0:13-28"
-    "###);
-
-    assert_yaml_snapshot!(parse_expr(r#"1.6..rel.col"#).unwrap(), @r###"
-    ---
-    Range:
-      start:
-        Literal:
-          Float: 1.6
-        span: "0:14-17"
-      end:
-        Indirection:
-          base:
-            Ident: rel
-            span: "0:19-22"
-          field:
-            Name: col
-        span: "0:22-26"
-    span: "0:13-28"
-    "###);
-
-    assert_yaml_snapshot!(parse_expr(r#"6.."#).unwrap(), @r###"
-    ---
-    Range:
-      start:
-        Literal:
-          Integer: 6
-        span: "0:14-15"
-      end: ~
-    span: "0:13-19"
-    "###);
-    assert_yaml_snapshot!(parse_expr(r#"..7"#).unwrap(), @r###"
-    ---
-    Range:
-      start: ~
-      end:
-        Literal:
-          Integer: 7
-        span: "0:16-17"
-    span: "0:13-19"
-    "###);
-
-    assert_yaml_snapshot!(parse_expr(r#".."#).unwrap(), @r###"
-    ---
-    Range:
-      start: ~
-      end: ~
-    span: "0:13-18"
-    "###);
-
-    assert_yaml_snapshot!(parse_expr(r#"@2020-01-01..@2021-01-01"#).unwrap(), @r###"
-    ---
-    Range:
-      start:
-        Literal:
-          Date: 2020-01-01
-        span: "0:14-25"
-      end:
-        Literal:
-          Date: 2021-01-01
-        span: "0:27-38"
-    span: "0:13-40"
-    "###);
-}
-
-#[test]
-fn test_basic_exprs() {
-    assert_yaml_snapshot!(parse_expr(r#"country == "USA""#).unwrap(), @r###"
-    ---
-    Binary:
-      left:
-        Ident: country
-        span: "0:14-21"
-      op: Eq
-      right:
-        Literal:
-          String: USA
-        span: "0:25-30"
-    span: "0:13-32"
-    "###);
-    assert_yaml_snapshot!(parse_expr("select {a, b, c}").unwrap(), @r###"
-    ---
-    FuncCall:
-      name:
-        Ident: select
-        span: "0:14-20"
-      args:
-        - Tuple:
-            - Ident: a
-              span: "0:22-23"
-            - Ident: b
-              span: "0:25-26"
-            - Ident: c
-              span: "0:28-29"
-          span: "0:21-30"
-    span: "0:13-32"
-    "###);
-    assert_yaml_snapshot!(parse_expr(
-            "group {title, country} (
-                aggregate {sum salary}
-            )"
-        ).unwrap(), @r###"
-    ---
-    FuncCall:
-      name:
-        Ident: group
-        span: "0:14-19"
-      args:
-        - Tuple:
-            - Ident: title
-              span: "0:21-26"
-            - Ident: country
-              span: "0:28-35"
-          span: "0:20-36"
-        - FuncCall:
-            name:
-              Ident: aggregate
-              span: "0:55-64"
-            args:
-              - Tuple:
-                  - FuncCall:
-                      name:
-                        Ident: sum
-                        span: "0:66-69"
-                      args:
-                        - Ident: salary
-                          span: "0:70-76"
-                    span: "0:66-76"
-                span: "0:65-77"
-          span: "0:55-77"
-    span: "0:13-93"
-    "###);
-    assert_yaml_snapshot!(parse_expr(
-            r#"    filter country == "USA""#
-        ).unwrap(), @r###"
-    ---
-    FuncCall:
-      name:
-        Ident: filter
-        span: "0:18-24"
-      args:
-        - Binary:
-            left:
-              Ident: country
-              span: "0:25-32"
-            op: Eq
-            right:
-              Literal:
-                String: USA
-              span: "0:36-41"
-          span: "0:25-41"
-    span: "0:13-43"
-    "###);
-    assert_yaml_snapshot!(parse_expr("{a, b, c,}").unwrap(), @r###"
-    ---
-    Tuple:
-      - Ident: a
-        span: "0:15-16"
-      - Ident: b
-        span: "0:18-19"
-      - Ident: c
-        span: "0:21-22"
-    span: "0:13-26"
-    "###);
-    assert_yaml_snapshot!(parse_expr(
-            r#"{
-  gross_salary = salary + payroll_tax,
-  gross_cost   = gross_salary + benefits_cost
-}"#
-        ).unwrap(), @r###"
-    ---
-    Tuple:
-      - Binary:
-          left:
-            Ident: salary
-            span: "0:33-39"
-          op: Add
-          right:
-            Ident: payroll_tax
-            span: "0:42-53"
-        span: "0:33-53"
-        alias: gross_salary
-      - Binary:
-          left:
-            Ident: gross_salary
-            span: "0:72-84"
-          op: Add
-          right:
-            Ident: benefits_cost
-            span: "0:87-100"
-        span: "0:72-100"
-        alias: gross_cost
-    span: "0:13-104"
-    "###);
-    // Currently not putting comments in our parse tree, so this is blank.
-    assert_yaml_snapshot!(parse_single(
-            r#"# this is a comment
-        select a"#
-        ).unwrap(), @r###"
-    ---
-    - VarDef:
-        kind: Main
-        name: main
-        value:
-          FuncCall:
-            name:
-              Ident: select
-              span: "0:28-34"
-            args:
-              - Ident: a
-                span: "0:35-36"
-          span: "0:28-36"
-      span: "0:28-36"
-    "###);
-    assert_yaml_snapshot!(parse_expr(
-            "join side:left country (id==employee_id)"
-        ).unwrap(), @r###"
-    ---
-    FuncCall:
-      name:
-        Ident: join
-        span: "0:14-18"
-      args:
-        - Ident: country
-          span: "0:29-36"
-        - Binary:
-            left:
-              Ident: id
-              span: "0:38-40"
-            op: Eq
-            right:
-              Ident: employee_id
-              span: "0:42-53"
-          span: "0:38-53"
-      named_args:
-        side:
-          Ident: left
-          span: "0:24-28"
-    span: "0:13-56"
-    "###);
-    assert_yaml_snapshot!(parse_expr("1  + 2").unwrap(), @r###"
-    ---
-    Binary:
-      left:
-        Literal:
-          Integer: 1
-        span: "0:14-15"
-      op: Add
-      right:
-        Literal:
-          Integer: 2
-        span: "0:19-20"
-    span: "0:13-22"
-    "###)
-}
-
-#[test]
-fn test_string() {
-    let double_quoted_ast = parse_expr(r#"" U S A ""#).unwrap();
-    assert_yaml_snapshot!(double_quoted_ast, @r###"
-    ---
-    Literal:
-      String: " U S A "
-    span: "0:13-25"
-    "###);
-
-    let single_quoted_ast = parse_expr(r#"' U S A '"#).unwrap();
-    assert_eq!(single_quoted_ast, double_quoted_ast);
-
-    // Single quotes within double quotes should produce a string containing
-    // the single quotes (and vice versa).
-    assert_yaml_snapshot!(parse_expr(r#""' U S A '""#).unwrap(), @r###"
-    ---
-    Literal:
-      String: "' U S A '"
-    span: "0:13-27"
-    "###);
-    assert_yaml_snapshot!(parse_expr(r#"'" U S A "'"#).unwrap(), @r###"
-    ---
-    Literal:
-      String: "\" U S A \""
-    span: "0:13-27"
-    "###);
-
-    parse_expr(r#"" U S A"#).unwrap_err();
-    parse_expr(r#"" U S A '"#).unwrap_err();
-
-    assert_yaml_snapshot!(parse_expr(r#"" \nU S A ""#).unwrap(), @r###"
-    ---
-    Literal:
-      String: " \nU S A "
-    span: "0:13-27"
-    "###);
-
-    assert_yaml_snapshot!(parse_expr(r#"r" \nU S A ""#).unwrap(), @r###"
-    ---
-    Literal:
-      String: " \\nU S A "
-    span: "0:13-28"
-    "###);
-
-    let multi_double = parse_expr(
-        r#""""
-''
-Canada
-"
-
-""""#,
-    )
-    .unwrap();
-    assert_yaml_snapshot!(multi_double, @r###"
-    ---
-    Literal:
-      String: "\n''\nCanada\n\"\n\n"
-    span: "0:13-36"
-    "###);
-
-    let multi_single = parse_expr(
-        r#"'''
-Canada
-"
-"""
-
-'''"#,
-    )
-    .unwrap();
-    assert_yaml_snapshot!(multi_single, @r###"
-    ---
-    Literal:
-      String: "\nCanada\n\"\n\"\"\"\n\n"
-    span: "0:13-37"
-    "###);
-
-    assert_yaml_snapshot!(
-          parse_expr("''").unwrap(),
-          @r###"
-    ---
-    Literal:
-      String: ""
-    span: "0:13-18"
-    "###);
-}
-
-#[test]
-fn test_s_string() {
-    assert_yaml_snapshot!(parse_expr(r#"s"SUM({col})""#).unwrap(), @r###"
-    ---
-    SString:
-      - String: SUM(
-      - Expr:
-          expr:
-            Ident: col
-            span: "0:21-24"
-          format: ~
-      - String: )
-    span: "0:13-29"
-    "###);
-    assert_yaml_snapshot!(parse_expr(r#"s"SUM({rel.`Col name`})""#).unwrap(), @r###"
-    ---
-    SString:
-      - String: SUM(
-      - Expr:
-          expr:
-            Indirection:
-              base:
-                Ident: rel
-                span: "0:21-24"
-              field:
-                Name: Col name
-            span: "0:25-35"
-          format: ~
-      - String: )
-    span: "0:13-40"
-    "###)
-}
-
-#[test]
-fn test_s_string_braces() {
-    assert_yaml_snapshot!(parse_expr(r#"s"{{?crystal_var}}""#).unwrap(), @r###"
-    ---
-    SString:
-      - String: "{?crystal_var}"
-    span: "0:13-35"
-    "###);
-    assert_yaml_snapshot!(parse_expr(r#"s"foo{{bar""#).unwrap(), @r###"
-    ---
-    SString:
-      - String: "foo{bar"
-    span: "0:13-27"
-    "###);
-    parse_expr(r#"s"foo{{bar}""#).unwrap_err();
-}
-
-#[test]
-fn test_tuple() {
-    assert_yaml_snapshot!(parse_expr(r#"{1 + 1, 2}"#).unwrap(), @r###"
-    ---
-    Tuple:
-      - Binary:
-          left:
-            Literal:
-              Integer: 1
-            span: "0:15-16"
-          op: Add
-          right:
-            Literal:
-              Integer: 1
-            span: "0:19-20"
-        span: "0:15-20"
-      - Literal:
-          Integer: 2
-        span: "0:22-23"
-    span: "0:13-26"
-    "###);
-    assert_yaml_snapshot!(parse_expr(r#"{1 + (f 1), 2}"#).unwrap(), @r###"
-    ---
-    Tuple:
-      - Binary:
-          left:
-            Literal:
-              Integer: 1
-            span: "0:15-16"
-          op: Add
-          right:
-            FuncCall:
-              name:
-                Ident: f
-                span: "0:20-21"
-              args:
-                - Literal:
-                    Integer: 1
-                  span: "0:22-23"
-            span: "0:20-23"
-        span: "0:15-24"
-      - Literal:
-          Integer: 2
-        span: "0:26-27"
-    span: "0:13-30"
-    "###);
-    // Line breaks
-    assert_yaml_snapshot!(parse_expr(
-            r#"{1,
-
-                2}"#
-        ).unwrap(), @r###"
-    ---
-    Tuple:
-      - Literal:
-          Integer: 1
-        span: "0:15-16"
-      - Literal:
-          Integer: 2
-        span: "0:35-36"
-    span: "0:13-39"
-    "###);
-    // Function call in a tuple
-    let ab = parse_expr(r#"{a b}"#).unwrap();
-    let a_comma_b = parse_expr(r#"{a, b}"#).unwrap();
-    assert_yaml_snapshot!(ab, @r###"
-    ---
-    Tuple:
-      - FuncCall:
-          name:
-            Ident: a
-            span: "0:15-16"
-          args:
-            - Ident: b
-              span: "0:17-18"
-        span: "0:15-18"
-    span: "0:13-21"
-    "###);
-    assert_yaml_snapshot!(a_comma_b, @r###"
-    ---
-    Tuple:
-      - Ident: a
-        span: "0:15-16"
-      - Ident: b
-        span: "0:18-19"
-    span: "0:13-22"
-    "###);
-    assert_ne!(ab, a_comma_b);
-
-    assert_yaml_snapshot!(parse_expr(r#"{amount, +amount, -amount}"#).unwrap(), @r###"
-    ---
-    Tuple:
-      - Ident: amount
-        span: "0:15-21"
-      - Unary:
-          op: Add
-          expr:
-            Ident: amount
-            span: "0:24-30"
-        span: "0:23-30"
-      - Unary:
-          op: Neg
-          expr:
-            Ident: amount
-            span: "0:33-39"
-        span: "0:32-39"
-    span: "0:13-42"
-    "###);
-    // Operators in tuple items
-    assert_yaml_snapshot!(parse_expr(r#"{amount, +amount, -amount}"#).unwrap(), @r###"
-    ---
-    Tuple:
-      - Ident: amount
-        span: "0:15-21"
-      - Unary:
-          op: Add
-          expr:
-            Ident: amount
-            span: "0:24-30"
-        span: "0:23-30"
-      - Unary:
-          op: Neg
-          expr:
-            Ident: amount
-            span: "0:33-39"
-        span: "0:32-39"
-    span: "0:13-42"
-    "###);
-}
-
-#[test]
-fn test_number() {
-    assert_yaml_snapshot!(parse_expr(r#"23"#).unwrap(), @r###"
-    ---
-    Literal:
-      Integer: 23
-    span: "0:13-18"
-    "###);
-    assert_yaml_snapshot!(parse_expr(r#"2_3_4.5_6"#).unwrap(), @r###"
-    ---
-    Literal:
-      Float: 234.56
-    span: "0:13-25"
-    "###);
-    assert_yaml_snapshot!(parse_expr(r#"23.6"#).unwrap(), @r###"
-    ---
-    Literal:
-      Float: 23.6
-    span: "0:13-20"
-    "###);
-    assert_yaml_snapshot!(parse_expr(r#"23.0"#).unwrap(), @r###"
-    ---
-    Literal:
-      Float: 23
-    span: "0:13-20"
-    "###);
-    assert_yaml_snapshot!(parse_expr(r#"2 + 2"#).unwrap(), @r###"
-    ---
-    Binary:
-      left:
-        Literal:
-          Integer: 2
-        span: "0:14-15"
-      op: Add
-      right:
-        Literal:
-          Integer: 2
-        span: "0:18-19"
-    span: "0:13-21"
-    "###);
-
-    // Underscores at the beginning are parsed as ident
-    assert!(parse_expr("_2").unwrap().kind.into_ident().is_ok());
-    assert!(parse_expr("_").unwrap().kind.into_ident().is_ok());
-
-    // We don't allow trailing periods
-    assert!(parse_expr(r#"add 1. (2, 3)"#).is_err());
-
-    assert!(parse_expr("_2.3").unwrap().kind.is_indirection());
-
-    assert_yaml_snapshot!(parse_expr(r#"2e3"#).unwrap(), @r###"
-    ---
-    Literal:
-      Float: 2000
-    span: "0:13-19"
-    "###);
-
-    // expr_of_string("2_").unwrap_err(); // TODO
-    // expr_of_string("2.3_").unwrap_err(); // TODO
-}
-
-#[test]
 fn test_filter() {
     assert_yaml_snapshot!(
             parse_single(r#"filter country == "USA""#).unwrap(), @r###"
@@ -1009,186 +344,26 @@ fn test_aggregate() {
 }
 
 #[test]
-fn test_derive() {
-    assert_yaml_snapshot!(
-            parse_expr(r#"derive {x = 5, y = (-x)}"#).unwrap()
-        , @r###"
+fn test_basic_exprs() {
+    // Currently not putting comments in our parse tree, so this is blank.
+    assert_yaml_snapshot!(parse_single(
+            r#"# this is a comment
+        select a"#
+        ).unwrap(), @r###"
     ---
-    FuncCall:
-      name:
-        Ident: derive
-        span: "0:14-20"
-      args:
-        - Tuple:
-            - Literal:
-                Integer: 5
-              span: "0:26-27"
-              alias: x
-            - Unary:
-                op: Neg
-                expr:
-                  Ident: x
-                  span: "0:35-36"
-              span: "0:33-37"
-              alias: y
-          span: "0:21-38"
-    span: "0:13-40"
-    "###);
-}
-
-#[test]
-fn test_select() {
-    assert_yaml_snapshot!(
-            parse_expr(r#"select x"#).unwrap()
-        , @r###"
-    ---
-    FuncCall:
-      name:
-        Ident: select
-        span: "0:14-20"
-      args:
-        - Ident: x
-          span: "0:21-22"
-    span: "0:13-24"
-    "###);
-
-    assert_yaml_snapshot!(
-            parse_expr(r#"select !{x}"#).unwrap()
-        , @r###"
-    ---
-    FuncCall:
-      name:
-        Ident: select
-        span: "0:14-20"
-      args:
-        - Unary:
-            op: Not
-            expr:
-              Tuple:
-                - Ident: x
-                  span: "0:23-24"
-              span: "0:22-25"
-          span: "0:21-25"
-    span: "0:13-27"
-    "###);
-
-    assert_yaml_snapshot!(
-            parse_expr(r#"select {x, y}"#).unwrap()
-        , @r###"
-    ---
-    FuncCall:
-      name:
-        Ident: select
-        span: "0:14-20"
-      args:
-        - Tuple:
-            - Ident: x
-              span: "0:22-23"
-            - Ident: y
-              span: "0:25-26"
-          span: "0:21-27"
-    span: "0:13-29"
-    "###);
-}
-
-#[test]
-fn test_expr() {
-    assert_yaml_snapshot!(
-            parse_expr(r#"country == "USA""#).unwrap()
-        , @r###"
-    ---
-    Binary:
-      left:
-        Ident: country
-        span: "0:14-21"
-      op: Eq
-      right:
-        Literal:
-          String: USA
-        span: "0:25-30"
-    span: "0:13-32"
-    "###);
-    assert_yaml_snapshot!(parse_expr(
-                r#"{
-  gross_salary = salary + payroll_tax,
-  gross_cost   = gross_salary + benefits_cost,
-}"#).unwrap(), @r###"
-    ---
-    Tuple:
-      - Binary:
-          left:
-            Ident: salary
-            span: "0:33-39"
-          op: Add
-          right:
-            Ident: payroll_tax
-            span: "0:42-53"
-        span: "0:33-53"
-        alias: gross_salary
-      - Binary:
-          left:
-            Ident: gross_salary
-            span: "0:72-84"
-          op: Add
-          right:
-            Ident: benefits_cost
-            span: "0:87-100"
-        span: "0:72-100"
-        alias: gross_cost
-    span: "0:13-105"
-    "###);
-    assert_yaml_snapshot!(
-            parse_expr(
-                "(salary + payroll_tax) * (1 + tax_rate)"
-            ).unwrap(),
-            @r###"
-    ---
-    Binary:
-      left:
-        Binary:
-          left:
-            Ident: salary
-            span: "0:15-21"
-          op: Add
-          right:
-            Ident: payroll_tax
-            span: "0:24-35"
-        span: "0:15-35"
-      op: Mul
-      right:
-        Binary:
-          left:
-            Literal:
-              Integer: 1
-            span: "0:40-41"
-          op: Add
-          right:
-            Ident: tax_rate
-            span: "0:44-52"
-        span: "0:40-52"
-    span: "0:13-55"
-    "###);
-}
-
-#[test]
-fn test_regex() {
-    assert_yaml_snapshot!(
-            parse_expr(
-                "'oba' ~= 'foobar'"
-            ).unwrap(),
-            @r###"
-    ---
-    Binary:
-      left:
-        Literal:
-          String: oba
-        span: "0:14-19"
-      op: RegexSearch
-      right:
-        Literal:
-          String: foobar
-        span: "0:23-31"
-    span: "0:13-33"
+    - VarDef:
+        kind: Main
+        name: main
+        value:
+          FuncCall:
+            name:
+              Ident: select
+              span: "0:28-34"
+            args:
+              - Ident: a
+                span: "0:35-36"
+          span: "0:28-36"
+      span: "0:28-36"
     "###);
 }
 
@@ -1494,299 +669,6 @@ fn test_function() {
 }
 
 #[test]
-fn test_func_call() {
-    // Function without argument
-    let ast = parse_expr(r#"count"#).unwrap();
-    let ident = ast.kind.into_ident().unwrap();
-    assert_yaml_snapshot!(
-            ident, @r###"
-    ---
-    count
-    "###);
-
-    let ast = parse_expr(r#"s 'foo'"#).unwrap();
-    assert_yaml_snapshot!(
-            ast, @r###"
-    ---
-    FuncCall:
-      name:
-        Ident: s
-        span: "0:14-15"
-      args:
-        - Literal:
-            String: foo
-          span: "0:16-21"
-    span: "0:13-23"
-    "###);
-
-    // A non-friendly option for #154
-    let ast = parse_expr(r#"count s'*'"#).unwrap();
-    let func_call: FuncCall = ast.kind.into_func_call().unwrap();
-    assert_yaml_snapshot!(
-            func_call, @r###"
-    ---
-    name:
-      Ident: count
-      span: "0:14-19"
-    args:
-      - SString:
-          - String: "*"
-        span: "0:20-24"
-    "###);
-
-    parse_expr("plus_one x:0 x:0 ").unwrap_err();
-
-    let ast = parse_expr(r#"add bar to=3"#).unwrap();
-    assert_yaml_snapshot!(
-            ast, @r###"
-    ---
-    FuncCall:
-      name:
-        Ident: add
-        span: "0:14-17"
-      args:
-        - Ident: bar
-          span: "0:18-21"
-        - Literal:
-            Integer: 3
-          span: "0:25-26"
-          alias: to
-    span: "0:13-28"
-    "###);
-}
-
-#[test]
-fn test_right_assoc() {
-    assert_yaml_snapshot!(parse_expr(r#"2 ** 3 ** 4"#).unwrap(), @r###"
-    ---
-    Binary:
-      left:
-        Literal:
-          Integer: 2
-        span: "0:14-15"
-      op: Pow
-      right:
-        Binary:
-          left:
-            Literal:
-              Integer: 3
-            span: "0:19-20"
-          op: Pow
-          right:
-            Literal:
-              Integer: 4
-            span: "0:24-25"
-        span: "0:19-25"
-    span: "0:13-27"
-    "###);
-    assert_yaml_snapshot!(parse_expr(r#"1 + 2 ** (3 + 4) ** 4"#).unwrap(), @r###"
-    ---
-    Binary:
-      left:
-        Literal:
-          Integer: 1
-        span: "0:14-15"
-      op: Add
-      right:
-        Binary:
-          left:
-            Literal:
-              Integer: 2
-            span: "0:18-19"
-          op: Pow
-          right:
-            Binary:
-              left:
-                Binary:
-                  left:
-                    Literal:
-                      Integer: 3
-                    span: "0:24-25"
-                  op: Add
-                  right:
-                    Literal:
-                      Integer: 4
-                    span: "0:28-29"
-                span: "0:24-29"
-              op: Pow
-              right:
-                Literal:
-                  Integer: 4
-                span: "0:34-35"
-            span: "0:23-35"
-        span: "0:18-35"
-    span: "0:13-37"
-    "###);
-}
-
-#[test]
-fn test_op_precedence() {
-    assert_yaml_snapshot!(parse_expr(r#"1 + 2 - 3 - 4"#).unwrap(), @r###"
-    ---
-    Binary:
-      left:
-        Binary:
-          left:
-            Binary:
-              left:
-                Literal:
-                  Integer: 1
-                span: "0:14-15"
-              op: Add
-              right:
-                Literal:
-                  Integer: 2
-                span: "0:18-19"
-            span: "0:14-19"
-          op: Sub
-          right:
-            Literal:
-              Integer: 3
-            span: "0:22-23"
-        span: "0:14-23"
-      op: Sub
-      right:
-        Literal:
-          Integer: 4
-        span: "0:26-27"
-    span: "0:13-29"
-    "###);
-
-    assert_yaml_snapshot!(parse_expr(r#"1 / (3 * 4)"#).unwrap(), @r###"
-    ---
-    Binary:
-      left:
-        Literal:
-          Integer: 1
-        span: "0:14-15"
-      op: DivFloat
-      right:
-        Binary:
-          left:
-            Literal:
-              Integer: 3
-            span: "0:19-20"
-          op: Mul
-          right:
-            Literal:
-              Integer: 4
-            span: "0:23-24"
-        span: "0:19-24"
-    span: "0:13-27"
-    "###);
-
-    assert_yaml_snapshot!(parse_expr(r#"1 / 2 - 3 * 4 + 1"#).unwrap(), @r###"
-    ---
-    Binary:
-      left:
-        Binary:
-          left:
-            Binary:
-              left:
-                Literal:
-                  Integer: 1
-                span: "0:14-15"
-              op: DivFloat
-              right:
-                Literal:
-                  Integer: 2
-                span: "0:18-19"
-            span: "0:14-19"
-          op: Sub
-          right:
-            Binary:
-              left:
-                Literal:
-                  Integer: 3
-                span: "0:22-23"
-              op: Mul
-              right:
-                Literal:
-                  Integer: 4
-                span: "0:26-27"
-            span: "0:22-27"
-        span: "0:14-27"
-      op: Add
-      right:
-        Literal:
-          Integer: 1
-        span: "0:30-31"
-    span: "0:13-33"
-    "###);
-
-    assert_yaml_snapshot!(parse_expr(r#"a && b || !c && d"#).unwrap(), @r###"
-    ---
-    Binary:
-      left:
-        Binary:
-          left:
-            Ident: a
-            span: "0:14-15"
-          op: And
-          right:
-            Ident: b
-            span: "0:19-20"
-        span: "0:14-20"
-      op: Or
-      right:
-        Binary:
-          left:
-            Unary:
-              op: Not
-              expr:
-                Ident: c
-                span: "0:25-26"
-            span: "0:24-26"
-          op: And
-          right:
-            Ident: d
-            span: "0:30-31"
-        span: "0:24-31"
-    span: "0:13-33"
-    "###);
-
-    assert_yaml_snapshot!(parse_expr(r#"a && b + c || (d e) && f"#).unwrap(), @r###"
-    ---
-    Binary:
-      left:
-        Binary:
-          left:
-            Ident: a
-            span: "0:14-15"
-          op: And
-          right:
-            Binary:
-              left:
-                Ident: b
-                span: "0:19-20"
-              op: Add
-              right:
-                Ident: c
-                span: "0:23-24"
-            span: "0:19-24"
-        span: "0:14-24"
-      op: Or
-      right:
-        Binary:
-          left:
-            FuncCall:
-              name:
-                Ident: d
-                span: "0:29-30"
-              args:
-                - Ident: e
-                  span: "0:31-32"
-            span: "0:29-32"
-          op: And
-          right:
-            Ident: f
-            span: "0:37-38"
-        span: "0:28-38"
-    span: "0:13-40"
-    "###);
-}
-
-#[test]
 fn test_var_def() {
     assert_yaml_snapshot!(parse_single(
             "let newest_employees = (from employees)"
@@ -1950,23 +832,6 @@ fn test_var_def() {
 
 #[test]
 fn test_inline_pipeline() {
-    assert_yaml_snapshot!(parse_expr("(salary | percentile 50)").unwrap(), @r###"
-    ---
-    Pipeline:
-      exprs:
-        - Ident: salary
-          span: "0:15-21"
-        - FuncCall:
-            name:
-              Ident: percentile
-              span: "0:24-34"
-            args:
-              - Literal:
-                  Integer: 50
-                span: "0:35-37"
-          span: "0:24-37"
-    span: "0:13-40"
-    "###);
     assert_yaml_snapshot!(parse_single("let median = x -> (x | percentile 50)\n").unwrap(), @r###"
     ---
     - VarDef:
@@ -2297,32 +1162,6 @@ fn test_dates() {
           span: "0:9-76"
       span: "0:9-77"
     "###);
-
-    assert_yaml_snapshot!(parse_expr("@2011-02-01").unwrap(), @r###"
-    ---
-    Literal:
-      Date: 2011-02-01
-    span: "0:13-27"
-    "###);
-    assert_yaml_snapshot!(parse_expr("@2011-02-01T10:00").unwrap(), @r###"
-    ---
-    Literal:
-      Timestamp: "2011-02-01T10:00"
-    span: "0:13-33"
-    "###);
-    assert_yaml_snapshot!(parse_expr("@14:00").unwrap(), @r###"
-    ---
-    Literal:
-      Time: "14:00"
-    span: "0:13-22"
-    "###);
-    // assert_yaml_snapshot!(parse_expr("@2011-02-01T10:00<datetime>").unwrap(), @"");
-
-    parse_expr("@2020-01-0").unwrap_err();
-
-    parse_expr("@2020-01-011").unwrap_err();
-
-    parse_expr("@2020-01-01T111").unwrap_err();
 }
 
 #[test]
@@ -2664,87 +1503,6 @@ join s=salaries (==id)
 }
 
 #[test]
-fn test_ident_with_keywords() {
-    assert_yaml_snapshot!(parse_expr(r"select {andrew, orion, lettuce, falsehood, null0}").unwrap(), @r###"
-    ---
-    FuncCall:
-      name:
-        Ident: select
-        span: "0:14-20"
-      args:
-        - Tuple:
-            - Ident: andrew
-              span: "0:22-28"
-            - Ident: orion
-              span: "0:30-35"
-            - Ident: lettuce
-              span: "0:37-44"
-            - Ident: falsehood
-              span: "0:46-55"
-            - Ident: null0
-              span: "0:57-62"
-          span: "0:21-63"
-    span: "0:13-65"
-    "###);
-
-    assert_yaml_snapshot!(parse_expr(r"{false}").unwrap(), @r###"
-    ---
-    Tuple:
-      - Literal:
-          Boolean: false
-        span: "0:15-20"
-    span: "0:13-23"
-    "###);
-}
-
-#[test]
-fn test_case() {
-    assert_yaml_snapshot!(parse_expr(r#"case [
-            nickname != null => nickname,
-            true => null
-        ]"#).unwrap(), @r###"
-    ---
-    Case:
-      - condition:
-          Binary:
-            left:
-              Ident: nickname
-              span: "0:33-41"
-            op: Ne
-            right:
-              Literal: "Null"
-              span: "0:45-49"
-          span: "0:33-49"
-        value:
-          Ident: nickname
-          span: "0:53-61"
-      - condition:
-          Literal:
-            Boolean: true
-          span: "0:75-79"
-        value:
-          Literal: "Null"
-          span: "0:83-87"
-    span: "0:13-99"
-    "###);
-}
-
-#[test]
-fn test_params() {
-    assert_yaml_snapshot!(parse_expr(r#"$2"#).unwrap(), @r###"
-    ---
-    Param: "2"
-    span: "0:13-18"
-    "###);
-
-    assert_yaml_snapshot!(parse_expr(r#"$2_any_text"#).unwrap(), @r###"
-    ---
-    Param: 2_any_text
-    span: "0:13-27"
-    "###);
-}
-
-#[test]
 fn test_unicode() {
     let source = "from tète";
     assert_yaml_snapshot!(parse_single(source).unwrap(), @r###"
@@ -3037,48 +1795,12 @@ fn test_module() {
 }
 
 #[test]
-fn test_lookup_01() {
-    assert_yaml_snapshot!(parse_expr(
-    r#"
-      {a = {x = 2}}.a.x
-    "#,
-    ).unwrap(), @r###"
-    ---
-    Indirection:
-      base:
-        Indirection:
-          base:
-            Tuple:
-              - Tuple:
-                  - Literal:
-                      Integer: 2
-                    span: "0:31-32"
-                    alias: x
-                span: "0:26-33"
-                alias: a
-            span: "0:21-34"
-          field:
-            Name: a
-        span: "0:34-36"
-      field:
-        Name: x
-    span: "0:13-45"
-    "###);
-}
-
-#[test]
-fn test_lookup_02() {
-    assert_yaml_snapshot!(parse_expr(
-    r#"
-      hello.*
-    "#,
-    ).unwrap(), @r###"
-    ---
-    Indirection:
-      base:
-        Ident: hello
-        span: "0:21-26"
-      field: Star
-    span: "0:13-35"
-    "###);
+fn test_number() {
+    // We don't allow trailing periods
+    assert!(parse_single(
+        r#"
+    from artists
+    derive x = 1."#
+    )
+    .is_err());
 }
